Added new option -xHost <fqdn hostname> this allows to specify a hostname on commandline to set the host=<fqdn> metapost variable. This helps to setup eg dyndns hostnames on the use for civservers. I don't like to use a dyndns domain in my local network ;-).
Side effect, if the option isn't given, the parameter host for the metapost 
will be setup one time in srv_init and not on each generate_metaserver_post call.

The size of srvarg.metasendhost of 256 charakters seems to long, but
RFC 2181 defines this length (see http://www.faqs.org/rfcs/rfc2181.html section 11).



git-svn-id: file:///usr/local/google/home/rocall/svn/freeciv-warclient/trunk@1508 4efe6140-7f1d-0410-9bd2-be1ee28408da